+++ package.json
@@ -12,6 +12,7 @@
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "preact": "^10.23.1",
+    "preact-router": "^4.1.2",
     "react-router-dom": "^6.25.1",
     "realm-web": "^2.0.1",
     "typescript": "^5.5.4"
