+++ src/Header.jsx
@@ -46,6 +46,7 @@ const Header = () => {
                             alt="Dark Mode"
                         />
                     </button>
+                    
                 </nav>
             </div>
         </div>
